[Refactor] 온보딩, 프로필 수정 뷰 공용화를 위한 UserProfileInputView 추가
